# backend/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "AreaPreviewRequest": {
        "properties": {
          "api_endpoint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Endpoint"
          },
          "network_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Network Id",
            "type": "string"
          }
        },
        "required": [
          "network_id"
        ],
        "title": "AreaPreviewRequest",
        "type": "object"
      },
...
      "ClliOut": {
        "properties": {
          "clli_code": {
            "title": "Clli Code",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "clli_code",
          "description"
        ],
        "title": "ClliOut",
        "type": "object"
      },
...
      "GrantOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "property_id": {
            "title": "Property Id",
            "type": "integer"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "user_id",
          "property_id",
          "created_at"
        ],
        "title": "GrantOut",
        "type": "object"
      },
      "GrantRequest": {
        "properties": {
          "property_id": {
            "title": "Property Id",
            "type": "integer"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "user_id",
          "property_id"
        ],
        "title": "GrantRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Common Area Monitor API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/access": {
      "delete": {
        "operationId": "revoke_access_api_v1_admin_access_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Access",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "grant_access_api_v1_admin_access_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Grant Access",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/areas/preview": {
      "post": {
        "description": "SPEC \u00a76.1 \u2014 live eero validation. Operator confirms the network_id\n+ name + eero count before saving.",
        "operationId": "preview_area_api_v1_admin_areas_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AreaPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AreaPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Area",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/areas/{area_id}": {
      "delete": {
        "operationId": "delete_common_area_api_v1_admin_areas__area_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "area_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Area Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Common Area",
        "tags": [
          "admin"
        ]
      },
      "put": {
        "operationId": "update_common_area_api_v1_admin_areas__area_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "area_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Area Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommonAreaUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonAreaOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Common Area",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/clli/olt": {
      "get": {
        "operationId": "list_olt_cllis_api_v1_admin_clli_olt_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClliOut"
                  },
                  "title": "Response List Olt Cllis Api V1 Admin Clli Olt Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Olt Cllis",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "create_olt_clli_api_v1_admin_clli_olt_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClliCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClliOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Olt Clli",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/clli/olt/{clli_id}": {
      "delete": {
        "operationId": "delete_olt_clli_api_v1_admin_clli_olt__clli_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "clli_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Clli Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Olt Clli",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/clli/seven-fifty": {
      "get": {
        "operationId": "list_seven_fifty_cllis_api_v1_admin_clli_seven_fifty_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClliOut"
                  },
                  "title": "Response List Seven Fifty Cllis Api V1 Admin Clli Seven Fifty Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Seven Fifty Cllis",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "create_seven_fifty_clli_api_v1_admin_clli_seven_fifty_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClliCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClliOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Seven Fifty Clli",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/clli/seven-fifty/{clli_id}": {
      "delete": {
        "operationId": "delete_seven_fifty_clli_api_v1_admin_clli_seven_fifty__clli_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "clli_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Clli Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Seven Fifty Clli",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/maintenance": {
      "post": {
        "operationId": "create_maintenance_api_v1_admin_maintenance_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Maintenance",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/maintenance/{m_id}": {
      "delete": {
        "operationId": "delete_maintenance_api_v1_admin_maintenance__m_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "m_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "M Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Maintenance",
        "tags": [
          "admin"
        ]
      },
      "put": {
        "operationId": "update_maintenance_api_v1_admin_maintenance__m_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "m_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "M Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Maintenance",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/properties": {
      "get": {
        "operationId": "list_properties_api_v1_admin_properties_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PropertyOut"
                  },
                  "title": "Response List Properties Api V1 Admin Properties Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Properties",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "create_property_api_v1_admin_properties_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PropertyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Property",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/properties/{property_id}": {
      "delete": {
        "operationId": "delete_property_api_v1_admin_properties__property_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "property_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Property Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Property",
        "tags": [
          "admin"
        ]
      },
      "put": {
        "operationId": "update_property_api_v1_admin_properties__property_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "property_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Property Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PropertyUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Property",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/properties/{property_id}/areas": {
      "post": {
        "operationId": "create_common_area_api_v1_admin_properties__property_id__areas_post",
        "parameters": [
          {
            "in": "path",
            "name": "property_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Property Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommonAreaCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonAreaOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Common Area",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/areas/{area_id}": {
      "get": {
        "description": "`area_id` is the network_id (e.g. `AKS-001`). Mock-mode short-circuits;\nreal-mode enforces per-property access via the parent property.",
        "operationId": "get_area_api_v1_areas__area_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "area_id",
            "required": true,
            "schema": {
              "title": "Area Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AreaDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Common-area detail (SPEC \u00a75.6)",
        "tags": [
          "dashboard"
        ]
      }
    },
...
  }
}